refactor(check): merge multiple files into one
