--- cmd/check/check_command.go
@@ -1,19 +1,0 @@
-package main
-
-import (
-	"time"
-
-	"github.com/henry40408/ssh-shell-resource/internal"
-)
-
-func CheckCommand(request *CheckRequest) CheckResponse {
-	versions := CheckResponse{}
-
-	previousVersion := request.Request.Version
-	if !previousVersion.Timestamp.IsZero() {
-		versions = append(versions, previousVersion)
-	}
-
-	versions = append(versions, internal.Version{Timestamp: time.Now()})
-	return versions
-}
