+++ cmd/check/main.go
@@ -3,10 +3,29 @@ package main
 import (
 	"fmt"
 	"os"
+	"time"
 
 	"github.com/henry40408/ssh-shell-resource/internal"
 	"github.com/spacemonkeygo/errors"
 )
+
+type CheckRequest struct {
+	internal.Request
+}
+
+type CheckResponse []internal.Version
+
+func CheckCommand(request *CheckRequest) CheckResponse {
+	versions := CheckResponse{}
+
+	previousVersion := request.Request.Version
+	if !previousVersion.Timestamp.IsZero() {
+		versions = append(versions, previousVersion)
+	}
+
+	versions = append(versions, internal.Version{Timestamp: time.Now()})
+	return versions
+}
 
 func Main(stdin, stdout *os.File) error {
 	var request CheckRequest
